+++ tests/errmsgs/tnim_echo_quitting.nim
@@ -1,9 +1,9 @@
 discard """
-  cmd: "nim check -d:nimEchoQuitting $file"
+  cmd: "nim check $file"
   action: "reject"
   nimout: '''
 tnim_echo_quitting.nim(12, 1) Error: undeclared identifier: 'nonexistant'
-nim quitting
+NimQuittingError
 '''
 """